+++ src/main/scala/com/kodekutters/neo4j/Neo4jLoader.scala
@@ -16,8 +16,7 @@ import scala.collection.JavaConverters._
   *
   * @author R. Wathelet June 2017
   *
-  * ref: https://github.com/workingDog/scalastix
-  *
+  *         ref: https://github.com/workingDog/scalastix
   * @param inFile the input file to process
   * @param dbDir  the neo4j graph database directory name
   */
@@ -37,6 +36,10 @@ class Neo4jLoader(inFile: String, dbDir: String) {
   def processBundle(bundle: Bundle) = {
     // all nodes and their internal relations are created first
     bundle.objects.foreach(nodesMaker.createNodes(_))
+    // print the number of SDO, SRO and StixObj (MarkingDefinition+LanguageContent)
+    nodesMaker.counter.foreach(n => println(n._1 + ": " + n._2))
+    // sum the SDO, SRO and StixObj
+    println("total: " + nodesMaker.counter.foldLeft(0)(_ + _._2))
     // all SRO and relations that depends on nodes are created after the nodes
     bundle.objects.foreach(relsMaker.createRelations(_))
   }